+++ StreamerMode/index.ts
@@ -114,7 +114,7 @@ function getBestLocalIPv4(): {
   );
 }
 
-const VERSION = "1.2.0";
+const VERSION = "1.2.1";
 const DEFAULT_PORT = 3959;
 
 type EffectResponseEntry = Omit<EffectEntry, "id"> & {